Changed few files

+++ features/tasks.py
@@ -1,12 +1,14 @@
 import sqlite3
 from features.utility import *
 
-def schedule_task(email):
-    task=input("Enter the task you want to schedule :- ")
-    deadline=input("Enter the deadline for the task (DD-MM-YYYY): ").strip()
+def schedule_task(email, task, deadline = None):
+    if not task:
+        return f"❌ Task description cannot be empty. Please try again."
+ 
+    if not deadline:
+        return f"❌ Deadline cannot be empty. Please try again."
     if not date_checker(deadline):
-        print("❌ Invalid date format. Please enter the date in DD-MM-YYYY format.")
-        return schedule_task(email)
+         return f"❌ Invalid date format. Please enter the date in DD-MM-YYYY format."
     conn=sqlite3.connect("student_toolkit.db")
     cursor=conn.cursor()
     cursor.execute('''
@@ -15,6 +17,7 @@ def schedule_task(email):
     print("✅ Task scheduled successfully!")
     conn.commit()
     conn.close()
+    return f"✅ Task scheduled successfully!"
 
 def view_tasks(email):
     conn=sqlite3.connect("student_toolkit.db")
@@ -22,52 +25,65 @@ def view_tasks(email):
     cursor.execute("SELECT * FROM tasks WHERE student_email = ?", (email,))
     tasks=cursor.fetchall()
     if tasks:
-        print(50*"=")
-        print("\t\t-- Tasks --")
-        for task in tasks:
-            print(f"Task: {task[2]}, Deadline: {task[4]}")
-        print(50*"=")
-    else:
-        print("❌ No tasks found.")
+         conn.commit()
+         conn.close()
+         return tasks
     conn.close()
+    return f"❌ No tasks found."
 
-def delete_tasks(email):
+def delete_tasks(email, task_id=None):
     conn=sqlite3.connect("student_toolkit.db")
     cursor=conn.cursor()
     cursor.execute("SELECT * FROM tasks WHERE student_email = ?", (email,))
     tasks=cursor.fetchall()
     if tasks:
-        
+         
         print("✅ Tasks found!")
-        
+        flag=False
         for task in tasks:
-            print(f"ID: {task[0]}, Task: {task[2]}, Deadline: {task[4]}")
-        task_id=input("Enter the ID of the task you want to delete: ")
-        cursor.execute("DELETE FROM tasks WHERE id = ? AND student_email = ?", (task_id, email))
-        conn.commit()
-        print("✅ Task deleted successfully.")
+            if task[0]==int(task_id):
+                flag=True
+                break
+        if flag:
+            cursor.execute("DELETE FROM tasks WHERE id = ? AND student_email = ?", (task_id, email))
+            conn.commit()
+            conn.close()
+            return f"✅ Task deleted successfully."
+        else:
+            conn.commit()
+            conn.close()
+            return f"❌ Wrong task ID. Please try again."
     else:
-        print("❌ No tasks found.")
-    conn.close()
-
-def Completed_tasks(email):
+        conn.commit()
+        conn.close()
+        return f"❌ No tasks found."
+ 
+def Completed_tasks(email, task_id=None):
     conn=sqlite3.connect("student_toolkit.db")
     cursor=conn.cursor()
     cursor.execute("SELECT * FROM tasks WHERE student_email = ? and completed = 0", (email,))
     tasks=cursor.fetchall()
     if tasks:
-        print("✅ Following are incomplete Tasks!")
+        flag=False
         for task in tasks:
-            print(f"ID: {task[0]}, Task: {task[2]}, Deadline: {task[4]}, Completed: {task[3]}")
-        task_id=input("Enter the ID of the task you want to mark as completed: ")
-        cursor.execute("UPDATE tasks SET completed = TRUE WHERE id = ? AND student_email = ?", (task_id, email))
-        conn.commit()
-        print("✅ Task is marked as completed.")
+            if task[0]==int(task_id):
+                flag=True
+                break
+        if flag:
+            cursor.execute("UPDATE tasks SET completed = TRUE WHERE id = ? AND student_email = ?", (task_id, email))
+            conn.commit()
+            conn.close()
+            return f"✅ Task is marked as completed."
+        else:
+            conn.commit()
+            conn.close()
+            return f"❌ Wrong task ID. Please try again."
     else:
-        print("❌ No tasks found.")
-    conn.close()
-
-def view_completed_tasks(email):
+        conn.commit()
+        conn.close()
+        return f"❌ No tasks found."
+ 
+def view_completed_tasks(email) :
     conn=sqlite3.connect("student_toolkit.db")
     cursor=conn.cursor()
     cursor.execute("SELECT * FROM tasks WHERE student_email = ? and completed = 1", (email,))
@@ -79,7 +95,7 @@ def view_completed_tasks(email):
             print(f"ID: {task[0]}, Task: {task[2]}, Deadline: {task[4]}, Completed: {task[3]}")
         print(50*"=")
     else:
-        print("❌ No completed tasks found.")
+        return f"❌ No completed tasks found."
     conn.close()
 
 def view_incomplete_tasks(email):
